+++ packages/front/src/lib/servicios/pagos.ts
@@ -1,5 +1,15 @@
 import type { PagosEstudianteResponse } from "$lib";
-import type { Pago } from "../../app";
+
+export interface PagoCreate {
+        student: string;
+        concept: string;
+        method: 'cash' | 'transfer' | 'point';
+        amount: number;
+        date: string;
+        exchange_rate: number;
+        reference?: string | null;
+        billetes?: { serial: string; monto: number }[];
+}
 
 const API = 'http://127.0.0.1:8000/api/pagos';
 
@@ -10,15 +20,15 @@ export const obtenerPagosPorCedula = async (fetch: typeof window.fetch, cedula: 
 	return data as PagosEstudianteResponse;
 };
 
-export const crearPago = async (fetch: typeof window.fetch, pago: Pago) => {
+export const crearPago = async (fetch: typeof window.fetch, pago: PagoCreate) => {
 	const res = await fetch(`${API}/add`, {
 		method: 'POST',
 		headers: { 'Content-Type': 'application/json' },
 		body: JSON.stringify(pago)
 	});
 	if (!res.ok) throw new Error('Error al crear el pago');
-	const data = await res.json();
-	return data.data as Pago;
+        const data = await res.json();
+        return data.data as { pago_id: number };
 };
 
 export const generarReporte = async (fetch: typeof window.fetch, params: string) => {
