+++ packages/front/src/routes/(sistemas)/caja/pagos/registrar/+page.server.ts
@@ -1,4 +1,4 @@
-import { crearPago, generarReporte, obtenerEstudiantes } from '$lib';
+import { crearPago, obtenerEstudiantes } from '$lib';
 import type { Actions, PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ fetch }) => {
@@ -13,12 +13,13 @@ export const actions: Actions = {
 		const student = formData.get('student') as string;
 		const concept = formData.get('payment-concept') as string;
 		const method = formData.get('payment-method') as 'cash' | 'transfer' | 'point';
-		const amount = formData.get('amount') as string;
+                const amount = formData.get('amount') as string;
+                const tasa = formData.get('tasa_divisa') as string;
 		const fecha_pago = formData.get('fecha_pago') as string;
 		const referencia = formData.get('referencia_transferencia') as string | null;
 		const billetesRaw = formData.getAll('billetes');
 
-		if (!student || !concept || !method || !amount || !fecha_pago) {
+                if (!student || !concept || !method || !amount || !fecha_pago || !tasa) {
 			return {
 				type: 'failure',
 				message: 'Por favor completa todos los campos'
@@ -37,15 +38,16 @@ export const actions: Actions = {
 			}
 		}
 
-		const payload = {
-			student,
-			concept,
-			method,
-			amount: parseFloat(amount),
-			date: fecha_pago,
-			reference: method === 'transfer' ? referencia : undefined,
-			billetes
-		};
+                const payload = {
+                        student,
+                        concept,
+                        method,
+                        amount: parseFloat(amount),
+                        date: fecha_pago,
+                        exchange_rate: parseFloat(tasa),
+                        reference: method === 'transfer' ? referencia : undefined,
+                        billetes
+                };
 
 		try {
 			await crearPago(fetch, payload);
@@ -62,38 +64,4 @@ export const actions: Actions = {
 			};
 		}
 	},
-	generarReporte: async ({ fetch, request }) => {
-		const formData = await request.formData();
-
-		const tipo = formData.get('tipo');
-		const filtro = formData.get('filtro');
-		const fecha = formData.get('fecha');
-		const fi = formData.get('fi');
-		const ff = formData.get('ff');
-
-		const params = new URLSearchParams();
-		if (tipo) params.set('tipo', tipo.toString());
-		if (filtro && filtro !== 'todos') params.set('f', filtro.toString());
-		if (tipo === 'dia' && fecha) params.set('d', fecha.toString());
-		if ((tipo === 'fechas' || tipo === 'monto') && fi && ff) {
-			params.set('fi', fi.toString());
-			params.set('ff', ff.toString());
-		}
-
-		try {
-			const { base64 } = await generarReporte(fetch, params.toString());
-			return {
-				base64,
-				type: 'application/pdf',
-				message: 'Reporte generado exitosamente',
-				invalidate: true
-			};
-		} catch (e: any) {
-			console.error('Error al generar reporte:', e);
-			return {
-				type: 'failure',
-				message: e.message
-			};
-		}
-	}
 };
